Restructure NGO and Participant Codebases
